LayoutTests/media/modern-media-controls/layout-node/node-made-dirty-during-layout-expected.txt

Testing the LayoutNode.y property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS node.element.style.left is "20px"
PASS node.element.style.left is "10px"
PASS successfullyParsed is true

TEST COMPLETE
